+++ core/auth/phone.go
@@ -3,6 +3,7 @@ package auth
 import (
 	"context"
 	"core-building-block/core/model"
+	"core-building-block/utils"
 	"encoding/base64"
 	"encoding/json"
 	"io/ioutil"
@@ -202,7 +203,7 @@ func (a *twilioPhoneAuthImpl) checkVerification(phone string, data url.Values, l
 		return errors.ErrorData(logutils.StatusInvalid, logutils.TypeString, &logutils.FieldArgs{"expected phone": phone, "actual phone": checkResponse.To})
 	}
 	if checkResponse.Status != "approved" {
-		return errors.ErrorData(logutils.StatusInvalid, typeVerificationStatus, &logutils.FieldArgs{"expected approved, actual:": checkResponse.Status})
+		return errors.ErrorData(logutils.StatusInvalid, typeVerificationStatus, &logutils.FieldArgs{"expected approved, actual:": checkResponse.Status}).SetStatus(utils.ErrorStatusInvalid)
 	}
 
 	return nil